Extracted GamesProvider to Interface
