--- src/main/java/org/gpc4j/ncaaf/views/CFB_View.java
@@ -1,14 +1,0 @@
-package org.gpc4j.ncaaf.views;
-
-/**
- *
- * @author Lyle T Harris
- */
-public class CFB_View extends NCAAF_View {
-
-    public CFB_View(int year) {
-        super("cfb.ftl",year);
-    }
-
-
-}
