+++ src/test/java/org/gpc4j/ncaaf/BowlGamesTest.java
@@ -4,6 +4,8 @@ import org.gpc4j.ncaaf.jaxb.Game;
 import org.gpc4j.ncaaf.jaxb.Team;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
+
+import org.gpc4j.ncaaf.redis.RedisGamesProvider;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.slf4j.LoggerFactory;
@@ -34,7 +36,7 @@ public class BowlGamesTest {
         //  pool = new JedisPool(cfg, "localhost", 6379, 0, "welcome1", 10, "JUnit");
         pool = new JedisPool(cfg, "macmini.local", 6388, 0, "welcome1", 10, "JUnit");
 
-        gp = new GamesProvider(pool);
+        gp = new RedisGamesProvider(pool);
         tp = new TeamProvider(pool);
     }
